:truck: moves donorModel into clinicalMongo
